pyproject.toml
[tool.poetry]
name = "ecomarketapi"
version = "0.1.0"
description = ""
authors = ["Akbota <aqbotamaratqyzy@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
psycopg = "^3.1.14"
drf-spectacular = "^0.27.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
